requirements/dev.txt
# Development dependencies
-r base.txt

# Type checking
mypy==1.11.2
types-flask==1.1.6

# Code quality
black==24.8.0
flake8==7.1.1
isort==5.13.2
pylint==3.2.7

# Testing (included from test.txt for convenience)
pytest==7.4.4
pytest-flask==1.3.0
pytest-mock==3.12.0
pytest-cov==5.0.0
flask-testing==0.8.1
